Leave/requirements.txt
pandas
python-docx
openpyxl
pyinstaller
